requirements.txt
aioredis==1.3.0
amqp==2.5.1
apipkg==1.5
apturl==0.5.2
asgi-redis==1.0.0
asgiref==3.2.2
asn1crypto==0.24.0
astroid==2.2.5
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
autobahn==19.9.3
Automat==0.7.0
awscli==1.16.198
beautifulsoup4==4.8.1
billiard==3.6.0.0
blinker==1.4
boto3==1.9.188
botocore==1.12.188
Brlapi==0.6.7
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
channels==2.3.0
channels-redis==2.4.0
chardet==3.0.4
Click==7.0
CodeConvert==3.0.2
colorama==0.3.9
command-not-found==0.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
crontab==0.22.6
cryptography==2.7
cssutils==1.0.2
cupshelpers==1.0
daphne==2.3.0
datadispatch==1.0.0
defer==1.0.6
defusedxml==0.6.0
distro==1.3.0
distro-info===0.18ubuntu0.18.10.2
dj-static==0.0.6
Django==2.2.6
django-admin==2.0.0
django-bootstrap3==11.1.0
django-braces==1.13.0
django-colorful==1.3
django-common-helpers==0.9.2
django-cors-headers==3.0.2
django-cron==0.5.1
django-crontab==0.7.1
django-elasticsearch-dsl==0.5.1
django-elasticsearch-dsl-drf==0.18
django-excel-response2==2.0.8
django-extensions==2.2.1
django-inlinecss==0.3.0
django-models-ext==1.1.9
django-nine==0.2.2
django-oauth-toolkit==1.2.0
django-oauth2-provider==0.2.6.1
django-redis==4.10.0
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.1.2
django-short-url==1.1.4
django-six==1.0.4
django-social-share==1.3.2
django-storages==1.7.1
djangorestframework==3.10.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
docutils==0.14
ecdsa==0.13.2
elasticsearch==6.4.0
elasticsearch-dsl==6.1.0
entrypoints==0.3
execnet==1.7.1
Faker==0.9.1
Flask==1.1.1
furl==2.1.0
future==0.17.1
gevent==1.4.0
geventhttpclient-wheels==1.3.1.dev1
greenlet==0.4.15
gunicorn==19.9.0
hiredis==1.0.0
httplib2==0.9.2
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.18
incremental==17.5.0
ipaddress==1.0.22
isort==4.3.21
itsdangerous==1.1.0
itypes==1.1.0
jeepney==0.4.1
Jinja2==2.10.1
jmespath==0.9.4
keyring==19.2.0
keyrings.alt==3.1
kombu==4.6.5
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.2
locustio==0.12.1
louis==3.6.0
macaroonbakery==1.1.4
Mako==1.0.7
MarkupSafe==1.1.1
mccabe==0.6.1
mixer==6.1.3
more-itertools==7.1.0
msgpack==0.6.1
msgpack-python==0.5.6
mysqlclient==1.4.6
netifaces==0.10.4
oauth==1.0.1
oauth2-provider==0.0
oauthlib==3.1.0
olefile==0.46
openapi-codec==1.3.2
orderedmultidict==1.0.1
pbr==5.4.4
pexpect==4.6.0
Pillow==5.2.0
protobuf==3.0.0
pyasn1==0.4.8
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pyee==6.0.0
PyGObject==3.30.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==0.9.3
PyNaCl==1.2.1
pynliner==0.8.0
pyRFC3339==1.1
python-apt==1.7.1
python-dateutil==2.8.1
python-debian==0.1.33
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2019.3
pyxdg==0.25
PyYAML==5.1
pyzmq==18.1.1
redis==3.3.11
reportlab==3.5.6
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rsa==3.4.2
s3transfer==0.2.1
screen==1.0.1
SecretStorage==3.1.1
shortuuid==0.5.0
simplejson==3.17.0
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.5
sqlparse==0.3.0
static3==0.7.0
system-service==0.3
systemd-python==234
testresources==2.0.1
text-unidecode==1.2
TimeConvert==1.5.0
Twisted==19.10.0
txaio==18.8.1
typed-ast==1.4.0
tzlocal==2.0.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.0
urllib3==1.25.7
usb-creator==0.3.3
validate-email==1.3
vine==1.3.0
virtualenv==16.7.8
wadllib==1.3.3
Werkzeug==0.16.0
wrapt==1.11.2
xkit==0.0.0
xlwt==1.3.0
zipp==0.6.0
zope.interface==4.7.1
